fix: PATRON2020-105 fix conflicts
