+++ frontend/src/components/List/List.jsx
@@ -10,7 +10,10 @@ const useStyles = makeStyles(theme => ({
     backgroundColor: '#e6e6e6',
     width: '100%',
     height: '100%',
-    alignContent: 'center'
+    alignContent: 'center',
+    margin: '0',
+    padding: '0',
+    boxSizing: 'border-box'
   }
 }))
