+++ frontend/src/views/__tests__/Layout.test.js
@@ -2,13 +2,22 @@
 
 import React from 'react'
 import { render, cleanup } from '@testing-library/react'
-import '@testing-library/jest-dom/extend-expect'
-import Dashboard from '../Dashboard.jsx'
+import Layout from '../Layout.jsx'
+import { MemoryRouter } from 'react-router-dom'
 
 describe('<DashboardContent />', () => {
+  let initialRoute
+  beforeEach(async () => {
+    initialRoute = ['/']
+  })
+
   afterEach(cleanup)
+
   it('renders DashboardContent component', async () => {
-    const { queryByTestId } = render(<Dashboard />)
+    const { queryByTestId } = render(
+      <MemoryRouter initialEntries={initialRoute}>
+        <Layout />
+      </MemoryRouter>)
     expect(queryByTestId('dashboard-id')).toBeTruthy()
   })
 })
